Set time output to verbose

+++ ruby/03_reference/projects/03_sudoku/lib/ui.rb
@@ -12,7 +12,7 @@ class UI
 
   def self.congratulate(game_start_time)
     elapsed_time = (Time.now - game_start_time).round
-    parse_time   = ChronicDuration.output(elapsed_time, format: :short)
+    parse_time   = ChronicDuration.output(elapsed_time, format: :long)
     congrats     = " Pretty good 👍\n Puzzle solved in #{parse_time}\n"
 
     puts Paint[congrats, 'gold', :bright]
